backend/requirements.txt
Django==3.2.11
djangorestframework==3.12.4
djangorestframework_simplejwt==5.0.0
itemadapter==0.4.0
pytz==2021.1
requests==2.27.0
Scrapy==2.5.1
SQLAlchemy==1.4.29
django_filter==21.1
python-dotenv==1.0.0
django-webpack-loader==1.4.1
psycopg2-binary==2.9.3
django-cors-headers==3.7.0
Twisted==22.10.0
pyopenssl==22.0.0
cryptography==38.0.4